requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
curl_cffi==0.13.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
multitasking==0.0.12
narwhals==2.8.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
peewee==3.18.2
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
requests-mock==1.12.1
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
yfinance==0.2.66
